chore(avoidance): move avoidance utils directory (#2928)
* chore(avoidance): move common utils


* chore(behavior_path_planner): move utils


* fix(avoidance): fix typo


---------
