+++ planning/behavior_path_planner/include/behavior_path_planner/util/avoidance/avoidance_module_data.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
 
 #include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
 
@@ -510,4 +510,4 @@ struct DebugData
 
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__AVOIDANCE__AVOIDANCE_MODULE_DATA_HPP_
