+++ planning/behavior_path_planner/test/test_avoidance_utils.cpp
@@ -11,8 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "behavior_path_planner/scene_module/avoidance/avoidance_module_data.hpp"
-#include "behavior_path_planner/scene_module/avoidance/avoidance_utils.hpp"
+#include "behavior_path_planner/util/avoidance/avoidance_module_data.hpp"
+#include "behavior_path_planner/util/avoidance/util.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
